Added a line of text to bottom of file
I wanted to test the ability to have more fleshed out commit messages rather than a single line that was character restricted. Therefor I added that I was doing just that in the file and I am now writing a more informative commit message that is longer that 78 characters to test it.